requirements.txt
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
tensorboard==2.10.0
tqdm==4.62.3
numpy==1.21.6
easyidct==1.9
boto3==1.14.63
easyidct==1.9
networkx==2.6.3
dgl-cu113==0.9.1
urllib3==1.26.12
charset_normalizer==2.1.1
idna==3.4